fix(codegen): define out-of-range casts and min literals
Float-to-integer `as` lowered to fptosi/fptoui, which yield poison for
any value the target cannot represent. The program's output stopped
being a function of its source: 1e300 as i32 printed -2147483648 at -O0,
printed nothing at all at -O3, and the constant folder computed a third
answer again. Lower through llvm.fptosi.sat / llvm.fptoui.sat instead,
and compute the same total function in FoldedConst::cast_to so a folded
cast and a run-time one agree.

The most negative value of a signed type had no literal spelling. A
negation is an operator over a literal, so the checker range-checked the
magnitude, which is one past the type's maximum; and the lexer carried
every literal in an i64, which excluded i64::MIN and the upper half of
u64 before a type was known. Integer tokens now carry an unsigned
magnitude, Literal::Integer carries an i128, and a negation over a
literal is checked against the negated value.

Four parser sign guards go with it as unreachable: no integer regex
admits a sign, so they could never fire.

+++ compiler/argument-binding/src/tests.rs
@@ -46,18 +46,18 @@ fn suppressed(name: &str) -> Parameter {
     }
 }
 
-fn int(value: i64) -> Expr {
+fn int(value: i128) -> Expr {
     Expr::Literal(Literal::Integer(value, None), span())
 }
 
 /// One argument as written: its optional call-site name and its value.
 type Arg = (Option<Identifier>, Expr);
 
-fn positional(value: i64) -> Arg {
+fn positional(value: i128) -> Arg {
     (None, int(value))
 }
 
-fn named(label: &str, value: i64) -> Arg {
+fn named(label: &str, value: i128) -> Arg {
     (Some(ident(label)), int(value))
 }
 
@@ -124,7 +124,7 @@ fn impl_block(type_name: &str, methods: Vec<MethodDef>) -> Item {
 }
 
 /// The single call in `main`'s body after binding, as the integer literals it holds.
-fn bound_values(items: &[Item]) -> Vec<i64> {
+fn bound_values(items: &[Item]) -> Vec<i128> {
     let Some(Item::Function(main)) = items.last() else {
         panic!("expected a trailing function item");
     };
@@ -412,7 +412,7 @@ fn a_call_nested_in_an_argument_is_bound_too() {
         panic!("expected a nested call");
     };
     assert!(arg_labels.is_empty(), "a label survived binding");
-    let values: Vec<i64> = inner
+    let values: Vec<i128> = inner
         .iter()
         .map(|a| match a {
             Expr::Literal(Literal::Integer(v, _), _) => *v,
@@ -591,7 +591,7 @@ fn a_call_nested_in_a_hoisted_argument_is_bound_too() {
         panic!("expected the nested call in the first temporary: {stmts:?}");
     };
     assert!(arg_labels.is_empty(), "a label survived binding");
-    let values: Vec<i64> = args
+    let values: Vec<i128> = args
         .iter()
         .map(|a| match a {
             Expr::Literal(Literal::Integer(v, _), _) => *v,
